+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Improve error message when it is not possible to access the data storage
+
 ## [1.5.0] - 2023-09-19
 
 ### Added